refactor(rest-api): extract legacy SSL normalizer to a shared package
