+++ gravitee-apim-rest-api/gravitee-apim-rest-api-service/src/main/java/io/gravitee/rest/api/service/common/LegacySslConfigurationNormalizer.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.gravitee.rest.api.service.v4.impl.validation;
+package io.gravitee.rest.api.service.common;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -21,13 +21,24 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import lombok.experimental.UtilityClass;
 
+/**
+ * Normalizes the legacy representation of the SSL "None" keyStore/trustStore discriminator.
+ * <p>
+ * Configurations written before gravitee-plugin-common-configurations 1.2.3 persist that
+ * discriminator as an empty string, while the shared SSL form schema now pins the "None"
+ * branch on {@code const: "NONE"}. The empty string therefore matches no {@code oneOf}
+ * branch, which makes both the generated form and the JSON schema validation reject it.
+ * <p>
+ * Both representations are equivalent for the gateway, which maps {@code ""} to
+ * {@code NONE} when deserializing.
+ */
 @UtilityClass
-class HttpProxySharedConfigurationNormalizer {
+public class LegacySslConfigurationNormalizer {
 
     private static final ObjectMapper OBJECT_MAPPER = new ObjectMapper();
     private static final String NONE = "NONE";
 
-    static String normalizeLegacySslNoneValues(String sharedConfiguration) {
+    public static String normalizeLegacySslNoneValues(String sharedConfiguration) {
         try {
             JsonNode root = OBJECT_MAPPER.readTree(sharedConfiguration);
             if (!(root.path("ssl") instanceof ObjectNode sslObject)) {
